Introduce `get_metric_or_err` helper and avoid spurious errors
We now avoid constructing an error when looking up the special
"comment" and "demonstration" metrics (since constructing an
error logs an error message).

In the future, we might want to consider using something like
`tracing-error` to capture spans, and defer the error logging
until we know that it's a "real" error.

Fixes #759