Require base URL in `get_chat_url` and related functions (#766)
All providers always passed this in as `Some`, except for the
'real' OpenAI provider (which relied on falling back to the default
OpenAI base url if the provider config didn't specify them).

To make sure that providers can't accidentally use the OpenAI base
url, I've changed the function to take in `&Url` instead of `Option<&Url`>.
The 'real' OpenAI provider now does the fallback behavior itself as needed.

Co-authored-by: Aaron Hill <aaron@tensorzero.com>