Use Arc<str> for model/provider name
This keeps the names cheap to clone/access, which avoiding the need
to have `'static` lifetimes for use with `tokio::spawn`. It's an
intermediate step towards removing all `Box::leak` calls.

Unfortunately, this creates a very large diff, mostly due to
the need to adjust tests to call '.into()' instead of '.to_string()
/.to_owned()' when constructing configs.

Most of the changes are purely mechanical, with a few exceptions:
* We now store `Arc<str>` instead of `String` in some places (e.g.
  `HashMap`s. This avoids the need to frequently convert between
  `String` and `Arc<str>`.
* Error messages still use `String`, so we convert from `Arc<str>`
  to `String` when an error occurs.
* We now use the `rc` feature of `Serde`, which allows us to deserialize
  directly to `Arc<str>`. This means that we should not rely on the
  results of `Arc` pointer comparisons across
  serialization/deserialization - however, we have no reason to use
  pointer comparisons, and the `Arc::eq` impl does a normal `PartialEq`
  comparison.
* Several structs no longer take lifetime parameters