Relax some 'static lifetime requirements
This moves us closer to being able to remove 'static from
&'static Config.

For `FunctionType`, I introduced a new `FunctionConfigType` enum
which just mirrors the enum variant (Chat/Json) without any data.
This is a little ugly, but it allows us to avoid holding a reference
to a `FunctionConfig` or introducing an `is_chat` boolean.

The `RetriesConfig` struct is just two fields, and can easily
be passed by value (and `Copy`) instead of a reference.

The combination of these two changes allows us to remove
all 'static lifetimes from `infer_model_request_stream`, which
unlocks further refactors.