Don't start the server when we have an invalid ClickHouse url (#792)
If we have observability enabled, but the ClickHouse url is
missing/malformed, then we don't start up the server (since we
will never be able to access ClickHouse). However, we still start
the server if the ClickHouse health check fails, since the gateway
and ClickHouse might be started simultaneously.

Fixes #790

Co-authored-by: Aaron Hill <aaron@tensorzero.com>
Co-authored-by: Viraj Mehta <virajmehta@users.noreply.github.com>