Explicitly set AWS bedrock region in e2e tensorzero.toml (#830)
The default provider in the AWS `RegionProviderChain` will try many
different strategies to detect the region, including making network
requests to '169.254.169.254' (which will time out when not running
on AWS).

We should change our behavior to avoid this footgun (e.g. require
an explicit opt-in for this network-based region detection behavior).
For now, let's speed up the test server by explicitly setting the
region, which will skip the region-detection logic