Add new `model_name` parameter to `/inference`
This parameter can be used instead of `function_name`.
When provided, we call a builtin "default" function, which simplify
calls the provided model without any additional configuration.
Only chat completions are supported, and templates/JSON schema
are not allowed.

Internally, we no longer store shorthand models in `ModelTable`.
`ModelTable.get` now creates shorthand models on-the-fly, which
allows uses to pass in shorthands with arbitrary model names
(e.g. `openai::gpt-9000`) to the model provider

All calls made using `model_name` are recorded in ClickHouse
under the function name `tensorzero::default`, with the variant
name set to the `model_name` provided by the user. This function
does not exist in the config - we dynamically create a `FunctionConfig`
for each request, with just a single variant for the `model_name`