Support shorthand model names for embedding models
This allows using model names like `openai::some-model-name` when
specifying any embedding model. This behaves in the same way as
shorthands for normal models - we'll create a model with the default
settings for the specified provider and model name.

Currently, this is only usable with `dicl` - you can now set
`embedding_model = "openai::text-embedding-3-small"` for a dicl
function.

Internally, we now have a generic `BaseModelTable` struct,
which handles the shorthand logic for both standard and
embedding models (with the model-specific behavior controlled through
a new `ShorthandModelConfig`).

We treat all provider names as reserved for both standard and embedding
models (e.g. you cannot declare an embedding model named
`vllm::fake_model`). However, the only embedding shorthand that works
in `embedding_model` is `openai::*` and `dummy::*` (since these are
the only two embedding providers that we currently support).