Add initial pyo3-based Python client (#833)
* Add initial pyo3-based Python client

This is a new Python client that wraps the rust client using pyo3.
We implement the streaming/non-streaming and sync/async apis supported
by the existing client, and all of the parameters for both /inference
and /feedback.

Constructing an embedded gateway is done through
`[Async]TensorZeroGateway.create_embedded_gateway`. This returns
a `Future` for the async gateway, and blocks for the sync gateway
(since we need to connect to ClickHouse and run the migrations).

The existing `test_client.py` was copied over almost verbatim
(I fixed a few test-case bugs that were expoesd by the stricter
serialization/deserialization logic in Rust). We now run all of
the tests twice - once in http gateway mode, and once in embedded
gateway mode.

Limitations:
* The error handling is fairly minimal - we throw a
  `TensorZeroInternalError` for most errors, which just provides
  a stringified error.
* We still implement the async/sync context managers on
  `[Async]TensorZeroGateway`, but it currently does nothing.
  Implementing this will be fairly tricky, since `reqwest` doesn't
  provide a way to close its connection pool.
* None of the code is particularly optimimized - it may hold the
  GIL for longer than needed, and incurs additional overhead from
  JSON serialization/deserialization between Python and Rust
* We create a fake `httpx.Response` object in `TensorZeroError`
  for backwards compatibility (since we now use `reqwest` from
  the rust code). The status code and message are available, but
  all other fields (e.g. response headers) are empty.
* We don't have types on any of the method parameters

This can coexist alongside the existing python client, until we decide
we're ready to replace the implementation.

* Run PyO3 client tests as part of merge queue

* Make test.sh use uv

* Make 'url' a workspace dep

* Use relative paths in test

* Start adding python-related comments

* Extract code to helper function

* Add link to pyo3 issue

* Address review comments

* Fix doc comment

* Format imports

* Fix another import error

* Address review comments

* Remove patchelf

* Fix github actions pytest

* Update Cargo.lock

* Update for 'function_name' being Option

* Add missing InferenceInput import

---------

Co-authored-by: Viraj Mehta <virajmehta@users.noreply.github.com>