Add 'json_mode' to `ChatCompletionInferenceParams`
This allows overriding the function json mode for both text
and json functions.

For text functions, this only passes
'json_mode' to the underlying model provider. We do *not* attempt
to perform any json or json-schema validation of the response
(not even to check if the response is valid json), so that users
are always able to get a response back from a text function.

This new functionality is used by the openai-compatible endpoint,
which passes through `response_format` to the real inference endpoint.