Make JsonMode optional in ChatCompletionConfig
This is preparation for supporting `json_mode` in chat variants.
For now, this does not result in any behavior changes - we use
unwrap it to `JsonMode::Strict` for json functions, and continue
to ignore `json_mode` for chat functions.

I've removed the `Default` impl, to make it harder to misuse this
(since the correct default will depend on the context)