Fix exception handling in `AsyncTensorZeroGateway.__aexit__`
We were accidentally returning `self` - this is a truthy value,
which signals that we want to suppress the exception that was
passed in as an argument.

We now return `()` (`None`), which matches the behavior of `__exit__`.
I've also added a test to ensure that errors correctly propagate through
our context managers.