Consistently use same "capital of Japan" prompt in batch/common tests (#974)
Our batch tests start new jobs with the prompt
"What is the capital city of Japan?", and check existing batch
jobs with `check_inference_params_response`. However, this function
was checking for a different prompt:
"What is the name of the capital city of Japan?"
which caused the batch tests to fail whenever an existing job
was polled.

We now use the prompt "What is the capital city of Japan?" everywhere.