Build python-pyo3 client for Python ABI 3.9
This allows us to build a single wheel (per platform)
which will work for Python 3.9 and above.

Some of the '&str'-based apis are not available on Python 3.9,
so I switched to using 'Cow<'_, str'>' apis instead.

I've also bumped the python-pyo3 version, so that we can deploy
to testpypi. This version will be replaced by a real version number
when we deploy to the real pypi