Add initial support for image input (#1081)
* Add initial support for image input

We now support passing in image content blocks. There are two
supported formats:

* URL: `{"type": "image", "url": "https://example.org/my-image.png"}`
* Base64: `{"type": "image", "mime_type": "image/png", "data": "<base64>"}

I've added support to the openai, google_ai_studio_gemini, and anthropic
providers.

When we receive an image url, we download it outselves, and then pass
the downloaded image to the model provider. This ensures that the
contents of our object store always matches the actual image
that we send to a model provider.

We write the image to the object store as part of the clickhouse
inference table writes. As a result, the `async_writes` also controls
whether or not we block the request on writing images to the object
store. However, an error when writing to the object store does not
cause the overall request to fail.

The object store path is currently `images/{hash}.{suffix}.base64`,
where `hash` is the BLAKE3 hash of the base64-encoded images, and
'suffix' is the filetype suffix (`png`, `jpg`, or `webp`)

The supported mime types are `image/png`, `image/jpg`, and `image/webp`
By default, any requests containing image content blocks will be
rejected. To allow image requests, you must configure an object store
in the gateway config:

```toml
[gateway.object_store]
type = "filesystem"
path = "/some/local/path"
```

The supported types are:
* `filesystem` - writes to a local directory
* `s3` - writes to an S3 bucket
* `disabled` - explicitly disables saving images.

When we serialize inputs and `raw_request` to ClickHouse, we strip
out the image content, to avoid filling up ClickHouse with images.

Main code changes:
* In addition to `InputMessageContent`, we now have
`ResolvedInputMessageContent`. An `InputMessageContent` represents the
content directly deserialized from the inference request, and can
include either an image or or a base64-encoded image. A
`ResolvedInputMessageContent` is produced after we download any image
urls, and only stores base64-encoded images. We now have `Resolved`
versions of other structs as well (e.g. `ResolvedInput`)
* A new `ObjectStoreData` struct manages the configured object store
 (wrapping the `object_store` crate).

* Fix clippy

* Fix batch tests

* Fix unused warning

* Fix batch/e2e feature compilation

* Make ObjectStoreData optional

* Add new error type

* Revert openai test changes

* Test that images are rejected without explicit object storage

* Fix clippy

* Fix merge fallout

* Add another allow(dead_code)

* Adjust base64

* Fix it

* Work on s3

* Get it running

* Get fs test working

* Adjust prints

* Fix anthropic image provider

* s/stripped/resolved/

* Rename ObjectStoreData to ObjectStoreInfo

* Work on review comments

* Split out image code and add unit test

* Adjust test

* Move to top-level `[object_storage]` config section

* Fix unit tests

* Fix batch tests warnings

* Use join_all for both clickhouse and object_store futures

* Move sanitize_raw_request to image.rs

* Add 'workflow_dispatch:' to merge queue checks

* [Temp] Add on-push to merge-queue

* Revert "[Temp] Add on-push to merge-queue"

This reverts commit 5b4a48ba99c90d0d054491ddfc406aca1ec0f72d.

* Store image path in 'input_messages'

* Run fmt

* Add test for fetching image from url

* Run fmt

* Fix a typo

* Remove unused 'typeid' dependency