Add support for 'extra_body' in model provider config (#1121)
* Add `ModelProvider` struct to wrap `ProviderConfig` enum

This will be used to hold generic model provider configuration
fields used by all provder implementations (e.g. an `extra_body`
field).

This PR is a pure refactoring, with no behavioral changes.

* Remove extra type hint

* Use serde(transparent)

* Add support for 'extra_body' in model provider config

This has a higher priority than the variant-level 'extra_body' -
if the same key is set in both, then the value from the model provider
config will be used in the constructed request.

* Fix batch tests

* Add comment and fix compilation

* Run fmt

* Fix batch tests

* Adjust tests

* Fix e2e tensorzero.toml for pointer format