Add 'POST /datasets/:dataset/datapoints' route to gateway
This is a port of the corresponding nodejs route in
'ui/app/routes/observability/inferences/$inference_id/route.tsx'

While implementing this route, I need to make a few other changes:
* 'Migration0014' no longer errors if 'ChatInferenceDataset' or
  'JsonInferenceDataset' are non-empty (since we're now inserting
  data into these tables from the e2e tests)
* 'deserialize_optional_json_string' now treats empty strings as
  'None'
* Several fields use 'deserialize_optional_json_string' or
  'deserialize_json_string' so that we can read them back
  from ClickHouse in the new endpoint.

I've opted not to add in retrying logic when fetching the
inference (or demonstration), under the assumption that users
won't try to create a datapoint immediately after running an inference
or providing a demonstration.

The tests are currently using 'chrono' to parse the 'updated_at'
timestamp. I entirely avoided timestamp handling in
the actual endpoint code, as we can just let ClickHouse fill
in the current time for 'updated_at' when inserting a row.