Always wrap inference clickhouse write in 'tokio::spawn'
This prevents the clickhouse write from getting cancelled,
even if our outer '/inference' future is cancelled.
We still await this task when async writes are disabled.

This reduces the chances that we perform a partial write to ClickHouse
(e.g. writing to 'ChatInference' but not 'ModelInference'), in
preparation for exposing a 'timeout' parameter on the embedded
gateway.