Add datapoint delete endpoint
The new endpoint is 'DELETE /datasets/:dataset/function/:function/kind/:kind/datapoint/:id'
Since this is an internal endpoint that will only be called from the ui
(for now), I've included all of the needed data as parameters, so that
we can efficiently insert into the correct table.

Deletion is implemented by inserting a new row into the ClickHouse
table with 'is_deleted' and 'updated_at' modified.