Implement caching for embedding requests
This is very similar to the existing non-streaming cache code.
The `EmbeddingModelConfig.embed` method now takes in
`InferenceClients`, which gives us access to the cache
settings.

For now, dicl is the only user of embeddings - the same
cache settings will be used for both the normal model
requests and the embedding request.