Avoid creating proxy-proxy-cache file for non-string body
When attempting to cache a non-string response in proxy-provider,
we would first create an empty file, and then skip writing to it
when we failed to process the server response. This broke some
of the image tests when running the provider-proxy in write mode,
since we run a small HTTP server which serves an image.

We now only open the proxy-proxy-cache file after successfully
serializing the response to a string, to reduce the chance that
we end up with an empty file.