Include raw_response for more streaming errors
When we fail to start reading fromthe stream due to a bad
status code, we try to read the response body (since it shouldn't
be an SEE stream) and use it as 'raw_response'. This makes errors
more useful when a streaming request fails due to invalid parameters
in the request.

We were previously doing this for OpenAI, but not for any other
providers.