Add `TensorZeroDeserialize` macro and use it to improve error paths (#1492)
* Add `TensorZeroDeserialize` macro and use it to improve error paths

This macro uses `serde_path_to_error` internally to work around
the issues caused by Serde's internal buffering with enums tagged
with `#[serde(tag = "...")]`

This macro is (mostly) a drop in-replacement for `serde::Deserialize`.
I've added it in a few places (in config parsing), along with a test
to show the nicer error message that we now produce.

* Fix review comments

* Fix newlines