Add 'extra_headers' config to variant/model-provider
For now, this is not exposed through inference.
The 'extra_headers' config value is a list of objects with
'name' and 'value' keys:

```
extra_headers = [
    { name = "anthropic-beta", value = "output-128k-2025-02-19"}
]
```

The variant-level 'extra_headers' is applied first, followed
by the model-provider 'extra_headers'. Entries in 'extra_headers'
can overwrite existing headers (whether set by TensorZero or
by a previous 'extra_headers' application).

This allows us to enable the 'anthropic-beta' header for
128K token output.