Introduce `Client[Input, InputMessage, InputMessageContent]` types (#1619)
* Introduce `Client[Input, InputMessage, InputMessageContent]` types

This allows us to take in the correct tool call type
(a new `ToolCallInput` type) for client inference requests.
This type has `arguments` as an `Option<Value>`, which allows us
to correctly serialize it as a JSON object when sending a request
to the server.

Previously, we were reusing the interal server `ToolCall` type, which
stored `arguments` as a `String` (constructed after handling
`arguments`/`raw_arguments` from the input). This prevented us
from ever serializing `arguments` as an object, which meant that we
would always trigger a deprecation warning on the server.

* Make deserialize_content private again