Remove Docker cache mount for 'target' directory (#1860)
In https://github.com/tensorzero/tensorzero/actions/runs/14524961439/job/40754404251?pr=1856
we observed a gateway binary **from another PR** being run.

Specifically, we got an error message:
"tensorzero::evaluator_inference_id is required when
tensorzero::datapoint_id and tensorzero::human_feedback are provided"

which was only added in a not-yet-merged PR:
https://github.com/tensorzero/tensorzero/pull/1855

This issue seems to be caused by the fact that Cargo
detects a stale crate by comparing file mtimes to
the dep-info mtime (which is located in the `target`
directory):
https://github.com/rust-lang/cargo/blob/185c6952617bcdb73d9082ad145d4f0b46a0dbfc/src/cargo/core/compiler/fingerprint/mod.rs#L29-L34

When we re-use a Docker cache volume for a 'target' directory,
Cargo will see that the dep-info has a newer mtime than our file
mtimes (since the dep-info mtime can come from a just-finished
build in another job), and skip rebuilding the crate.
However, this does not depend on the actual file contents, so
we can end up with an incorrect artifact (from a random open PR)