Only run Python/node tests in the live-tests job
We already run linters/typechecking in the 'validate' job
(which runs in both PR CI and the merge queue), so there's
no need to also run them here. We only need to run the tests
(which need credentials).

I've also removed some of the explicit 'uv venv' steps, as
'uv run' takes care of this automatically