Add initial OpenTelemetry OTLP export to gateway (#1971)
* Add initial OpenTelemetry OTLP export to gateway

This feature is off by default, and must be explicitly enabled
via a newly-added `export` config:

```toml
[gateway.export.otlp.traces]
enabled = true
```

The actual export url is configured via the standard OpenTelemetry
environment variables (e.g. `OTEL_EXPORTER_OTLP_TRACES_ENDPOINT`).

For now, we don't allow configuring OTLP export in the Python client
(this is a global setting, so it will be tricky to make it work with
multiple independent embedded gateways).

When enabled, we create OpenTelemetry spans for several HTTP endpoints:
* /inference
* /batch_inference/*
* /openai/v1/chat/completions
* /feedback

For inference endpoints, we emit several nested spans corresponding
to the inference request flow:
* `tensorzero_function_call` - covers the overall function call, and
  is emitted for both `/inference` and `/openai/v1/chat/completions`
  This allows consumers to easily configure monitoring/alerts for
  all function calls, regardless of which inference endpoint they use
  * `tensorzero_variant_infer` - covers each variant inference, nested
    under the 'tensorzero_function_call' span
    * `tensorzero_model_infer` - covers the model inference request.
      * `tensorzero_model_provider_request` - covers the actual HTTP
	request to the model providers. This has standard OpenTelemetry
        GenAI attributes set on the span.

We only emit specifically-taggged spans (currently, any span with an
`http.` or `otel.` tracing-level field name). This ensures that we don't
accidentally start exporting internal spans to customers (which we may
want to break over time as the codebase changes).

During e2e tests, we now start a Jaeger instance, and verify that we can
query an exported span after submitting an inference request to the
gateway. Other tests install a custom in-process OTEL span exporter,
and verify that we export the expected spans (and that the total count
matches exactly).

Future extensions:
* Add support for OTEL metric export
* Attach more attributes to our exported spans (error information, more
  GenAI attributes, etc.)

* Add missing file

* Fix tests

* Fix unit tests

* Rename otel spans

* Rename to 'function_inference'

* Fix span name