Always overwrite existing headers with extra-headers
In some providers, we were calling '.header' with our headers
(e.g. 'Authorization') after setting the user-supplied extra-headers
with '.headers'. This could result in duplicate headers getting sent
(since '.header' will append to any existing headers)

We now call '.headers' as the last method call before making the request
- this causes extra-headers to completely replace any existing headers
with the same name.

To catch this kind of issue, I've modified provider-proxy to produce a
failure response when duplicate headers are present in the request.