Fix OTEL export of top-level span and trace headers (#2058)
* Fix OTEL export of top-level span and trace headers

The HTTP span was previously getting created at level `trace`,
which we compiled out in release builds. We now use the
`tracing-opentelemetry-instrumentation-sdk` feature `tracing_level_info`
to ensure that we emit this span in release builds.

I've also fixed the tracing HTTP header propagation
(`traceparent` and `tracestate` headers) - due to a bug in
`tracing-opentelemetry`, our use of a reloadable layer was
preventing the HTTP layer from seeing the OTEL context:
https://github.com/tokio-rs/tracing-opentelemetry/issues/121

* Run setup_logs in Tokio runtime

* Run fmt