Rename internal `Image` content block to `File`
We still need to be able to deserialize old 'image' content blocks
from the database, so we now have `#[serde(alias = "image")]` on some
fields. The UI server ClickHouse code also handles deserializing "image"
content blocks.

In the frontend codebase, we now display an error when trying to render
a file with a mime-type that does not start with `image/`

This is preparation for supporting additional file types (e.g. PDF)