Correctly use downloaded gateway and ui images on CI
We were missing an 'image:' field in our docker-compose.yml files.
As a result, the images loaded by 'docker load' were being ignored in
favor of whatever image happened to be in the Namespace cache

We now build the image with a specific sha tag, and use that tag when
starting the containers. This should result in an error if the image
export ever breaks, rather than silently falling back to the latest
published image