Support 'application/pdf' file input and add support to OpenAI provider (#2285)
* Rename internal `Image` content block to `File`

We still need to be able to deserialize old 'image' content blocks
from the database, so we now have `#[serde(alias = "image")]` on some
fields. The UI server ClickHouse code also handles deserializing "image"
content blocks.

In the frontend codebase, we now display an error when trying to render
a file with a mime-type that does not start with `image/`

This is preparation for supporting additional file types (e.g. PDF)

* Rename ClientInputMessageContent::Image and fix build

* Add test for new db format

* Adjust expected uuid

* More ui test fixes

* Add new fixtures

* Add another fixture

* Support 'file' in python sft

* Rerun

* Test deserializing 'file' content block

* Fix clippy

* Fix openrouter

* Set explicit image

* Fix another image

* Rename module from image to file

* Add issue link

* Fix imports

* Run fmt

* Change expected db read from 'image' to 'file'

* Support 'application/pdf' file input and add support to OpenAI provider

We now store files in 'observability/files/' in the object store
provider, rather than 'observability/images'. The full path is stored
in the databsae, so we can still read existing images saved under
'observability/images'

* Change batch test tag due to change in expected output

* Regenerate model inference cache

* Regenerate again