Allow arbitrary mime-types for 'file' input
Unfortunately, the OpenAI apis require us to convert
between mime-types and file extensions (when guessing the
mime-type for the openai-compatible endpoint, and when
generating a file name when invoking an openai provider).

When the mime-type is not a 'known good' mime type
(current 'image/webp', 'image/png', 'image/jpeg', or
'application/pdf'), we log a warning when we attempt to
convert between mime-types/extensions in either direction.