Support no-schema templates with 'system_text/assistant_text/user_text'
When a variant has a template but no schema, we now provide a single
key based on the template kind: 'system_text', 'assistant_text',
or 'user_text'. The template will be evaluated for each text content
block (as well as the system message), and requires the text content
to be a string (not using the 'arguments' key).