Add `[timeouts]` section to model provider config (#2364)
* Add `[timeouts]` section to model provider config

This is the first step of adding timeout support to
variants, models, and model providers. The config looks like this:

```toml
[models.slow_with_timeout.providers.slow.timeouts]
non_streaming = { total_ms = 400 }
streaming = { ttft_ms = 500 }
```

All of these fields are optional, and default to no timeout
(the existing behavior before this PR)

`non_streaming.total_ms` sets a timeout for the entire HTTP
request/response for non-streaming inferences, while
'streaming.ttft_ms' sets a timeout for the time to first token.

If either of these errors is hit, we treat it as an error from
the model provider (like a normal HTTP error produced by the
model provider). Currently, models always move on to the next
provider (without retrying) when a provider produces any error.

* Run fmt