Add `[timeouts]` section to variant config
Most of this PR just refactors our variant-handling code to
add a wrapper 'VariantInfo' struct, which holds both a 'VariantConfig'
and the 'TimeoutsConfig'. The `Variant` trait is now implemented on
`VariantInfo`, which calls the wrapped `VariantInfo` with the timeout
applied.

The config section looks like this:

```toml
[functions.basic_test_variant_timeout.variants.slow_timeout.timeouts]
non_streaming = { total_ms = 400 }
streaming = { ttft_ms = 500 }
```

The `timeouts` values are the same as for model provider timeouts.