Add `discard_unknown_chunks` to model-provider and implement for Gemini
When this option (off-by-default) is enabled, we emit a warning
and discard unknown chunks that occur in a provider stream. The
default behavior is to produce an error in the stream.

This allows users to continue using providers that unexpectely
start returning chunk types that we don't yet support (e.g. when
Gemini started producing thought chunks). Over time, we'll want
to add support for this to all providers - only GCP Vertex Gemini
is supported by this PR