Load TensorZeroError/TensorZeroInternalError early on (#2487)
We need to construct an instance of these exception classes if
an error occurs in our `#[pymodule]` function (e.g. an invalid
`RUST_LOG` environment variable). Let's load these up-front, to make
sure that we don't cause a nested error when trying to load an exception
(in case the exception class cannot be loaded for some reason)