Allow OTLP layer creation to fail when trace export is disabled
This allows the gateway to start up (with a warning logged)
if we have an invalid `OTEL_` environment variable set,
as long as `[export.otlp.traces.enabled]` is false.

The embedded client (and the Python package) never actually enable
the OTLP layer, so we just ignore any error produced when we try
to construct the otel layer.