Add 'delete = true' option to extra_body config
This allows deleting an key the body, rather than replacing or inserting
a key.

The new type of `extra_body` entry looks like:
```
{ pointer = "/some/pointer, delete = true }
```

'delete' must be set to 'true' - setting `delete = false` produces
an error during config loading.

If the specified pointer does not exist in the object (due to a parent
being missing, an array index being out-of-bounds, or attempting
to index an array with a string), we log a warning and continue on.
This allows `delete = true` to be used to ensure that a particular key
does not exist in the body, without breaking the model provider if it
already does not exist