Use Mixtral model for groq e2e (#2622)
* Use Mixtral model for groq e2e

This improves our test pass rate (when running without a provider-proxy
cache) - the tool-calling tests now pass with only one or two retries.
Our rate limit on Groq is very low, so I've increased the total number
of retries and the backoff time to make sure that all of the tests
eventually make a non-rate-limited request.

I also removed several of the 'max_tokens' checks from the e2e tests.
This is just checking that we forward through 'max_tokens' from the
config - we already test it in lots of places, and checking it
everywere makes it more difficult to adjust the 'max_tokens' for
a variant when switching models.

* Fix overrides