Use stricter validation for OpenAI-compatible content blocks
Whenver we encountered a parse error, we would wrap the entire content
block in 'tensorzero::arguments', and produce a 'text' TensorZero
content block. This could result in confusing behavior when the user
wasn't trying to use a text block (e.g. if they used an unknown content
block type, or left off fields).

We now look for a 'tensorzero::arguments' or 'type' field - if either
of those are set we encounter a parse error, we assume that the user
was *not* trying to use the deprecated behavior (passing a json object
directly to the arguments of a TensorZero function).