Add streaming support to mixture-of-n variant (#2495)
* Add streaming support to mixture-of-n variant

When we have multiple candidates and the fuser successfully starts a
stream, this just returns the fuser stream from `infer_stream`. However,
if the fuser fails (or we only have one candidate), we need to turn
a non-streaming candidate response into a fake stream response.
We do this by mapping each content block into a new chunk, where
the chunk contains the entire contents of the content block.

In streaming mode, we add the candidate usages to the usage chunk
in the stream, or create a new usage chunk if the fuser stream doesn't
contain one

In order to test the usage behavior, I've adjusted some of the dummy
providers to produce usage information based on the number of chunks
that they output, rather than a fixed value

* Collect extra_usage chunk

* Re-add check

* Fix review comments

* Fix usage behavior

* More work

* More work

* Fix rebase fallout

* Fix some usage counts

* Fix openai-compat tests

* More work

* More work

* Fix python tests

* Adjust openai tests

* Fix node openai tests

* Fix golang tests

* Format

* Fix more tests