Handle 'thoughtSignature' in Google AI Studio Gemini
This provider can produce messages with a 'thoughtSignature' field,
which should be passed back to the model in multi-turn conversations.

Unfortunately, a 'thoughtSignature' can be associated with any 'part'
(which maps to our `ContentBlock`), while the TensorZero 'Thought'
only support a text thought with a signature.

To handle this case without changing our data model, we split a single
GCP part with a `thoughtSignature` into two output TensorZero content
blocks:
```
{"type": "thought", "text": null, "signature": "thoughtSignature_contents"}
{"type": "some_content_block", ... }
```

When we receive a `Thought` with an attached signature as input in
Google AI Studio, we require it to be followed by a non-Thought
content block. These two content blocks are merged into a single GCP
part, with the signature from the first block, and the contents
from the second block. If a thought has 'text' set, we warn and
discard it.

This only affects Google AI Studio, as GCP Vertex never emits
'thoughtSignature' (even for a request and model identical to
the Google AI Studio input)

This adds new dependencies between entries in the content blocks array,
which is unfortunate. However, valid 'thoughtSignatures' can only
be produced as part of a GCP model output, and cannot be constructed by
hand. Any user code that manually inserts/removes Thought blocks is already
semantically incorrect, so it should be okay for us to produce explicit
errors in this case.