Add '_internal_provider_type' field to Thought/ThoughtChunk (#2808)
* Add '_internal_provider_type' field to Thought/ThoughtChunk

This is set on all Thought/ThoughtChunks provided by TensorZero.
When we see an '_internal_provider_type' field in an input
'Thought', we discard the thought if it doesn't match the target
provider type (similar to how we handle 'unknown' content blocks)

This field is optional - when unset, we always pass the thought
to the model provider. This will probably not be meaningful if
the thought was produced by manually, or if it came from a different
provider.

* Add comment and regen bindings

* Fix pyo3 tests

* Reformat