Add `TomlRelativePath` wrapper type for config paths (#2890)
* Add `TomlRelativePath` wrapper type for config paths

We now use this type everywhere that we deserialize a `PathBuf` from
a toml file, and require it in methods that read from the filesystem.

This PR is a pure refactor. By restricting the ways that a
`TomlRelativePath` can be constructed and accessed, we'll be
able to enforce the correct toml-relative path behavior for
globbed configs (in a future PR)

* Run fmt