Limit concurrency for evaluations/inferences in ui e2e tests
When regenerating the fixtures, we need to run with an evaluations
concurrency limit of 1, and only load one of the (duplicate) image
inferences on the playground page. This prevents us from having multiple
in-flight inferences at once, which can cause cache entries to get
overwritten.

While we might be able to increase the concurrency limit in 'normal'
(non-regen) mode, we've had a lot of issues with this test. For now,
let's have regen and non-regen modes run exactly the same logic.