Introduce a new `GatewayHandle` wrapper type
This type is not cloneable, and wraps `AppStateData` (which remains
cloneable). When we add support for ClickHouse write batching, the
Drop impl for `GatewayHandle` will block and wait for the batcher
to exit, to ensure that we don't lose writes during a graceful shutdown.

See the comments on `GatewayHandle` for more details