Drain channel after close in ClickHouse batching implementation (#3379)
* Drain channel after close in ClickHouse batching implementation

The channel can still have messages when it's closed (e.g. if
the last ClickhouseConnectionInfo writes lots of messages and is
then immediately dropped). We were assuming that the channel was empty
when it was closed, which could cause us to drop messages during
shutdown.

This showed as up flakes in our evaluations tests with batch writes
enabled (since we check that all of the expected rows were written)

* Fix condition and add test

* Fix test

* Also drop batch_sender

* Fix clippy