Add embedding examples for azure and ollama (#3351)
* docs(embedding examples): add examples for azure and ollama
addresses #3083 and #3232
the openai/ example was updated to the newest tensorzero::embeddings_model syntax, openai and tensorzero dependencies were updated, and a requirements.txt file which was referenced in the README.md was also created.
added examples for azure and ollama, with both containing a tensorzero.toml config file, and the ollama example folder addtitionally containing a docker deployment
ollama example was tested, but azure example could not be tested.

* Fix name

* Cleanup

* Cleanup

* Cleanup

* Cleanup

* Fix uv

---------

Co-authored-by: Gabriel Bianconi <GabrielBianconi@users.noreply.github.com>
Co-authored-by: Gabriel Bianconi <1275491+GabrielBianconi@users.noreply.github.com>