Remove 'extra_handle' parameter from inference methods (#3477)
This only use of this method was to try to prevent deadlocks in
Python. We now do this by explicitly holding on to gateway handles
in our 'smart iterator' Python objects (StreamWrapper and
AsyncStreamWrapper), so we no longer need this.

Additionally, we currently block external users from enabling
batching with the python embedded client, so any batching-related
changes cannot currently affect Python users.