Introduce `LazyResolvedInput` and use it during inference
This type sits in between `Input` and `ResolvedInput` -
it holds normalized legacy input data (e.g.
`{"type": "text", "value": {}}` has been converted
to a template invocation), but file url inputs
have not yet been resolved to in-memory files.

This PR is a pure refactor, and is the first step
towards adding support for forwarding image urls
directly to model providers (without needing to
wait for the file to be downloaded by the gateway).

A follow-up PR will build on this to add a
similar `LazyRequestMessage/ResolvedRequestMessage`
split, which which will be used to allow model providers
to avoid taking in a fully-resolved image (which still
being able to `.await` the file data if the backend does
not have file url forwarding enabled).