Add unit tests for convergence of estimated optimal probabilities.
As samples means and variances converge, the estimated optimal probabilities
should converge to the true optimum. This convergence may be nonmonotone in
the convergence of the sample statistics due to the nonlinear optimization
problem which is sensitive to the ordering of the sample means, so we average
over multiple random runs to yield monotonicity with high probability.