Consolidate span creating and OTLP headers into a single middleware (#3801)
Previously, custom OTLP header extraction was separated from the
creation of the top-level request span - we used both a `TraceLayer`
and a custom middleware function.

We now register a single middleware function, which performs all of the
necessarily logic. This removes the need to store data in
`request.extesnions` - we can now just use normal function calls.