Improve shutdown handling of custom OpenTelemetry tracer (#3805)
* Consolidate span creating and OTLP headers into a single middleware

Previously, custom OTLP header extraction was separated from the
creation of the top-level request span - we used both a `TraceLayer`
and a custom middleware function.

We now register a single middleware function, which performs all of the
necessarily logic. This removes the need to store data in
`request.extesnions` - we can now just use normal function calls.

* Improve shutdown handling of custom tracer

The shutdown handling is now done in `CustomTracer::drop`.
Almost all of the logic around custom tracers is now called from
'tensorzero_tracing_middleware' (including the possible creation
of a fresh `CustomTracer`).

The 'build_with_context' is now very simple - it just looks up
an entry in the context, and delegates `build_with_context` on
an already-constructed `SdkTracer`.

By performing the shutdown logic in `Drop`, we ensure that we only run
the shutdown after all associated `Spans` have already been dropped,
avoiding the problem of shutting down the tracer before we've finished
executing all of our spans (e.g. if we create a span for the
rate-limiting `return_tickets` call, which happens in the background)

This also eliminates the need for a moka eviction handler, since
dropping the final `Arc<CustomTracer>` from anywhere is now enough
to perform a correct shutdown.

* Update comment