Add support for encrypted reasoning to OpenAI responses
This needs to be explicitly enabled at the model provider level
via a new 'include_encrypted_reasoning' flag, since only certain
models support it (OpenAI will error if you try to request
'reasoning.encrypted_content' for an unsupported model).

For now, this is only supported in non-streaming mode, and only
for outputs. There seems to be a bug in the responses API for streaming
mode, and supporting multi-turn encrypted thoughts will require us to
change our Thought content block